+++ pkg/scheduler/inspector_routers.go
@@ -9,7 +9,6 @@ import (
 	"github.com/google/uuid"
 	coreV1 "k8s.io/api/core/v1"
 	"k8s.io/klog/v2"
-
 	"volcano.sh/volcano/pkg/inspector"
 )
